+++ FlightPlanProject/FlightPlanPlugin.cs
@@ -1,7 +1,7 @@
 ﻿using BepInEx;
 using BepInEx.Configuration;
 using BepInEx.Logging;
-using FlightPlan.UI;
+
 using FPUtilities;
 using HarmonyLib;
 using KSP.Game;
@@ -17,6 +17,9 @@ using SpaceWarp.API.UI.Appbar;
 using System.Collections;
 using System.Reflection;
 using UnityEngine;
+
+using FlightPlan.KTools.UI;
+using FlightPlan.KTools;
 
 namespace FlightPlan;
 
@@ -160,7 +163,7 @@ public class FlightPlanPlugin : BaseSpaceWarpPlugin
     {
         base.OnInitialized();
 
-        FPSettings.Init(SettingsPath);
+        GeneralSettings.Init(SettingsPath);
 
         Instance = this;
 
@@ -245,8 +248,8 @@ public class FlightPlanPlugin : BaseSpaceWarpPlugin
 
     void save_rect_pos()
     {
-        FPSettings.window_x_pos = (int)windowRect.xMin;
-        FPSettings.window_y_pos = (int)windowRect.yMin;
+        GeneralSettings.window_x_pos = (int)windowRect.xMin;
+        GeneralSettings.window_y_pos = (int)windowRect.yMin;
     }
 
     /// <summary>
@@ -274,8 +277,8 @@ public class FlightPlanPlugin : BaseSpaceWarpPlugin
         if (interfaceEnabled && GUIenabled && activeVessel != null)
         {
             FPStyles.Init();
-            FlightPlan.UI.UIWindow.check_main_window_pos(ref windowRect);
-            GUI.skin = FPStyles.skin;
+            WindowTool.check_main_window_pos(ref windowRect);
+            GUI.skin = GenericStyle.skin;
 
             windowRect = GUILayout.Window(
                 GUIUtility.GetControlID(FocusType.Passive),
@@ -307,10 +310,10 @@ public class FlightPlanPlugin : BaseSpaceWarpPlugin
     private void FillWindow(int windowID)
     {
         TopButtons.Init(windowRect.width);
-        if ( TopButtons.IconButton(FPStyles.cross))
+        if ( TopButtons.Button(GenericStyle.cross))
             CloseWindow();
 
-        GUI.Label(new Rect(9, 2, 29, 29), FPStyles.icon, FPStyles.icons_label);
+        GUI.Label(new Rect(9, 2, 29, 29), GenericStyle.icon, GenericStyle.icons_label);
         
         if (selectingBody)
         {
@@ -785,7 +788,7 @@ public class FlightPlanPlugin : BaseSpaceWarpPlugin
 
     private void DrawSectionHeader(string sectionName, string value = "", float labelWidth = -1, GUIStyle valueStyle = null) // was (string sectionName, ref bool isPopout, string value = "")
     {
-        if (valueStyle == null) valueStyle = FPStyles.label;
+        if (valueStyle == null) valueStyle = GenericStyle.label;
         GUILayout.BeginHorizontal();
         // Don't need popout buttons for ROC
         // isPopout = isPopout ? !CloseButton() : UI_Tools.SmallButton("⇖", popoutBtnStyle);
@@ -830,7 +833,7 @@ public class FlightPlanPlugin : BaseSpaceWarpPlugin
 
     private void DrawButton(string buttonStr, ref bool button)
     {
-        button = UI_Tools.Button(buttonStr);
+        button = UI_Tools.SmallButton(buttonStr);
     }
 
     private void DrawToggleButton(string runString, ref bool button, string stopString = "")
@@ -858,7 +861,7 @@ public class FlightPlanPlugin : BaseSpaceWarpPlugin
     private double DrawButtonWithTextField(string entryName, ref bool button, double value, string unit = "")
     {
         GUILayout.BeginHorizontal();
-        button = UI_Tools.Button(entryName);
+        button = UI_Tools.SmallButton(entryName);
         GUILayout.Space(10);
 
         value = UI_Fields.DoubleField(entryName, value);
