+++ FlightPlanProject/KTools/UI/AssetsLoader.cs
@@ -1,0 +1,28 @@
+
+using UnityEngine;
+using SpaceWarp.API.Assets;
+
+namespace FlightPlan.KTools.UI;
+
+public class AssetsLoader
+{
+    public static Texture2D loadIcon(string path)
+    {
+        var imageTexture = AssetManager.GetAsset<Texture2D>($"flight_plan/images/{path}.png");
+
+        //   Check if the texture is null
+        if (imageTexture == null)
+        {
+            // Print an error message to the Console
+            Debug.LogError("Failed to load image texture from path: " + path);
+
+            // Print the full path of the resource
+            Debug.Log("Full resource path: " + Application.dataPath + "/" + path);
+
+            // Print the type of resource that was expected
+            Debug.Log("Expected resource type: Texture2D");
+        }
+
+        return imageTexture;
+    }
+}
